Ópera/requirements.txt
dash==3.2.0
dash_bootstrap_components==2.0.3
dash_bootstrap_templates==2.1.0
pandas==2.3.1
plotly==6.2.0
python-dotenv==1.1.1
supabase==2.18.0
